DLWGMMIV

Setup

DLWGMMIV can be installed from the package manager in Julia's REPL:

] add https://github.com/kgovernor/DLWGMMIV.jl.git

De Loecker and Warzynski GMM IV (DLWGMMIV) Production Function Estimation

DLWGMMIV.dlwGMMIV — Function

dlwGMMIV(year, plantid, Q, inputs...; <keyword arguments>)

Returns results for estimated betas from GMM IV for industry's production function prodF given panel data with year, plantid, inputs... (with independent/dynamic inputs listed first), and output, Q.

Results are returned as a NamedTuple.

Arguments

num_indp_inputs::Integer=1: the number of independent/dynamic inputs.

bstart::Vector{Float64}: the starting beta values for optimization. Defaults to zeros.

δ_nm::Float64=0.1: the Nelder-Mead shrink step parameter, delta.

max_iters::Integer=500: the maximum number of iterations ran in optimization.

Configurable Options

prodF::String: the production function to estimate. Default is "CD", Cobb-Douglas; other options include "tl", TransLog.

opt::String: the optimization method to use. Default is "nm", Nelder-Mead; other options include "LBFGS".

Examples

julia> using DLWGMMIV

julia> df = DLWGMMIV.sim_data(20000, 10, input_names = ["K", "L"], prod_params = [0.1, 0.25]).df;
[...]

julia> data = [df.time, df.firm, df.Y, df.K, df.L];

julia> res = dlwGMMIV(data...);

julia> println("
Converge = $(res.conv_msg)\n
Objective Value = $(res.other_results.crit)\n
betas = $(res.beta_dlw)
")
Converge = true

Objective Value = 1.074768000861536e-9

betas = (beta_x1 = 0.10230132568110045, beta_x2 = 0.26251910427769287)

julia> res = dlwGMMIV(data..., opt = "LBFGS");

julia> println("
Converge = $(res.conv_msg)\n
Objective Value = $(res.other_results.crit)\n
betas = $(res.beta_dlw)
")
Converge = true

Objective Value = 3.6354547893764317e-19

betas = (beta_x1 = 0.10230132472854156, beta_x2 = 0.2625190996231563)

julia> res = dlwGMMIV(data..., opt = "LBFGS", prodF = "tl");

julia> println("
Converge = $(res.conv_msg)\n
Objective Value = $(res.other_results.crit)\n
betas = $(res.beta_dlw)
")
Converge = true

Objective Value = 3.1089429872638027e-15

betas = (beta_x1 = 0.11394854106849515, beta_x2 = 1.9222327345314836, beta_x1x2 = -0.5175358152780514, beta_x12 = -0.04266689698177511, beta_x22 = 0.00703394572086843)

Simulating Firm Data

DLWGMMIV.sim_data — Function

DLWGMMIV.sim_data(N, T; <keyword arguments>)

Returns a NamedTuple containing a DataFrame of a panel dataset of N firms over T+1 periods with specified parameters.

Arguments

num_inputs::Integer=2: the total number of production inputs to generate.

num_indp_inputs::Integer=1: the number of independent inputs to generate.

input_names::Vector{String}: a list of input names. Default is `["K","L"]. Additional inputs get a value of "X1","X2",... .

prod_params::Vector{Real}: a list of parameters for the production function. Default is [0.1, 0.25]. Additional inputs get a value that is equal to 1 minus sum(prod_params) divided by the number of additional inputs, and TransLog second order terms get a value of 0.

cost_params::Vector{Real}: a list of parameters for the cost function. Default is [0, 0.15]. Additional inputs get a value of 0.

omega_params::Vector{Real}: a list of parameters for production technology function. Default is [0, 0.8, 0, 0].

indp_inputs_params:: Vector{Real}: a list of parameters for independent inputs process. Default is [1]. Additional independent inputs get a value of 1.

σ_ω::Real=1: the variance associated with the productivity shock each period.

indp_inputs_lnmean::Vector{Real}: a list of natural log of mean values for each independent input. Default is [5]. Additional independent inputs get a value of 5.

indp_inputs_lnvariance::Vector{Real}: a list of variances for each natural log of independent input. Default is [1]. Additional independent inputs get a value of 1.

seed::Integer: sets a seed for Random number generator. Default is -1, no seed set.

X_start::Integer=1000: set starting values for optimizer which calculates optimal level of dependent inputs for each firm.

Configurable Options

prodF::String: the production function parameter. Default is "CD", Cobb-Douglas; other options include "tl", TransLog.

costF::String: the cost function parameter. Default is "ce", constant elasticity.

Examples

julia> using DLWGMMIV

julia> df = DLWGMMIV.sim_data(20, 10).df
Sim Data for 2 inputs, CD

K Parameters:
    K_prod_params = 0.1 | K_cost_params = 0.0
L Parameters:
    L_prod_params = 0.25 | L_cost_params = 0.15

    First order derivative at optimal L is approximately zero: true

    Second order derivative at optimal L check: true

=======================

SUMMARY:
        100.0% of observations passed first order conditions.
        100.0% of observations passed second order conditions.

=======================
    
220×18 DataFrame
Row │ time   firm   S           Y           P          TC        omega_i    XI          K         L          C_K       C_L        rent_K   rent_L ⋯
    │ Int64  Int64  Float64     Float64     Float64    Float64   Float64    Float64     Float64   Any        Float64   Float64    Float64  Float6 ⋯
────┼──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
  1 │     0      1   0.594649    3.60282    -647.634   651.237    0.687066   0.0        650.454   0.808584   650.454   0.783221       1.0  0.9686 ⋯
  2 │     0      2   0.151929    1.02321    -119.13    120.154   -0.12898    0.0        119.931   0.270617   119.931   0.222438       1.0  0.8219  
  3 │     0      3  -0.0790732   0.4036      -89.6328   90.0364  -0.828257   0.0         89.9487  0.120514    89.9487  0.0877392      1.0  0.7280  
 ⋮  │    ⋮      ⋮        ⋮           ⋮           ⋮          ⋮          ⋮          ⋮          ⋮          ⋮         ⋮          ⋮             ⋮        ⋮   ⋱ 
219 │    10     19  -0.26828     0.120173   -188.737   188.857   -1.85054    0.0152269  188.831   0.0420262  188.831   0.0261245      1.0  0.6216  
220 │    10     20  -0.475736    0.0675116   -83.0301   83.0976  -2.21972   -1.24907     83.0829  0.025454    83.0829  0.0146764      1.0  0.5765 ⋯                                                                                                                      
                                                                                                                      5 columns and 215 rows omitted

julia> df = DLWGMMIV.sim_data(20, 10, num_inputs = 3, input_names = ["k", "l", "m"], prod_params = [0.1, 0.25, 0.2, 0.05], prodF ="tl").df
Sim Data for 3 inputs, tl

k Parameters:
    k_prod_params = 0.1 | k_cost_params = 0.0  
l Parameters:
    l_prod_params = 0.25 | l_cost_params = 0.15
m Parameters:
    m_prod_params = 0.2 | m_cost_params = 0.0  
kl Parameters:
    kl_prod_params = 0.05 |
km Parameters:
    km_prod_params = 0.0 |
lm Parameters:
    lm_prod_params = 0.0 |
k2 Parameters:
    k2_prod_params = 0.0 |
l2 Parameters:
    l2_prod_params = 0.0 |
m2 Parameters:
    m2_prod_params = 0.0 |

******************************************************************************
This program contains Ipopt, a library for large-scale nonlinear optimization.
    Ipopt is released as open source code under the Eclipse Public License (EPL).
            For more information visit https://github.com/coin-or/Ipopt
******************************************************************************

    First order derivative at optimal L is approximately zero: true

    Second order derivative at optimal L check: true

=======================

SUMMARY:
        100.0% of observations passed optimization generating the simulated data.
        100.0% of observations passed first order conditions.
        100.0% of observations passed second order conditions.

=======================

220×24 DataFrame
Row │ time   firm   XI         k          l          m           Y          S          P          TC        omega_i    termination     C_k        ⋯   
    │ Int64  Int64  Float64    Float64    Float64    Float64     Float64    Float64    Float64    Float64   Float64    String          Float64    ⋯
────┼────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
  1 │     0      1   0.0       245.378    0.0790147  0.0236492   0.118246   -1.53147   -245.337   245.456   -0.603513  LOCALLY_SOLVED  245.378    ⋯ 
  2 │     0      2   0.0       194.093    1.1433     0.522569    2.61285     0.465793  -193.17    195.782    0.494648  LOCALLY_SOLVED  194.093     
  3 │     0      3   0.0        55.9995   0.0584379  0.0194533   0.0972665  -1.66692    -55.9599   56.0572  -0.663383  LOCALLY_SOLVED   55.9995    
 ⋮  │     ⋮      ⋮        ⋮          ⋮          ⋮          ⋮           ⋮          ⋮            ⋮         ⋮          ⋮             ⋮             ⋮      ⋱ 
219 │    10     19   0.117099  118.572    1.55853    0.783866    3.91933     0.64574   -117.102   121.021    0.72018   LOCALLY_SOLVED  118.572     
220 │    10     20   0.460817    6.10519  0.227785   0.123259    0.616295   -0.741436    -5.7946    6.4109   0.257407  LOCALLY_SOLVED    6.10519  ⋯                                                                                                                     
                                                                                                                     11 columns and 215 rows omitted                                                                                                                      

Optimization Problem

Let $N_{independent}$ and $N$ be equal to num_indp_inputs and num_inputs, respectively. Let $X_t = \{x_{1,t}, \dots, x_{N,t}\} \in \mathbb{R}_+^N$ and $X_{t,independent} = \{x_{1,t}, \dots, x_{N_{independent},t}\}$. $N_{independent}$ must be greater than zero and less than $N$, $0 < N_{independent} < N$.

The firm chooses $X_{t,dependent} = X_t \setminus X_{t,independent}$ that optimizes profits in each period $t$, $t \in T = \{0,1,2,\dots\}$.

\[\begin{equation} \max_{X_{t,dependent}} Profit_t(X_t, \omega_t) \quad \forall t \in T \end{equation}\]

where $\omega_t$ is the technology parameter, and $Profit_t(X_t, \omega_t) = Output_t(X_t, \omega_t) - Cost_t(X_t)$. The technology parameter $\omega_t$ is subject to a Total Factor Productivity (TFP) shock $\xi_t$, where $\xi_t \sim \mathcal{N}(0, \sigma_{\omega})$.

Exogenous variables at period $t=0$ are given by a random process. Specifically, for $x_{i,0} \in X_{0,independent}$, $x_{i,0} \sim \mathcal{N}(\mu_{x_i}, \sigma_{x_i})$ and $\omega_0 = \xi_0$, where $\mu_{x_i}$ and $\sigma_{x_i}$ equal indp_inputs_lnmean[i] and indp_inputs_lnvariance[i]. For $t>0$,

\[x_{i,t} = \alpha_{x_i} x_{i,t-1}\]

\[\omega_t = \sum_{i=0}^{3}c_i\omega_{t-1}^i + \xi_t\]

where $\alpha_{x_i} =$ $\sigma_{\alpha} \times$indp_inputs_params[i], $\sigma_{\alpha} \sim \mathcal{N}(0, 1)$, and $c_i =$ omega_params[i].

Let $X_{t}^*$ be the optimal solution to (1). Then,

\[\begin{aligned} Y_t &= Output_t(X_t^*, \omega_t) \\ TC_t &= Cost_t(X_t^*) \\ P_t &= Y - TC \\ S_t &= log(Y_t) - \omega_t \\ C_{x_i,t} &= Cost_t(x_{i,t}^*) \\ rent_{x_i,t} &= C_{x_i,t} / x_{i,t}^* \\ share\_TC_{x_i,t} &= C_{x_i,t} / TC_t \\ share\_Y_{x_i,t} &= C_{x_i,t} / Y_t \\ \end{aligned}\]

Production Functions

Cobb Douglas

\[Output_t(X_t, \omega_t) = \exp(\omega_t)\prod_{i=1}^{N}x_{i,t}^{\beta_i}\]

where prod_params = $\{\beta_1, \beta_2, \dots, \beta_N\}$.

TransLog

\[\begin{align*} Output_t(X_t, \omega_t) &= \exp(\omega_t + \sum_{i=1}^{N}\beta_ilog(x_{i,t}) \\ &+ \sum_{i=1}^{N}\sum_{j=i}^{N}\beta_{i,j}log(x_{i,t})log(x_{j,t})) \end{align*}\]

where prod_params = $\{\beta_1, \beta_2, \dots, \beta_N,$ $\beta_{1,2}, \dots, \beta_{1,N}, \dots, \beta_{i,i+1}, \dots, \beta_{i,N}, \dots, \beta_{N-1,N},$ $\beta_{1,1}, \beta_{2,2}, \dots, \beta_{N,N}\}$.

Cost Functions

Constant Elasticity

\[Cost_t(X_t) = \sum_{i=1}^{N}x_{i,t}^{\gamma_i+1}\]

where cost_params[i] $= \gamma_i$.

Examples

Cobb Douglas Example

Let's first simulate data for industry A with 20,000 firms over 10 periods with an additional initial period 0. The production function in industry A is Cobb-Douglas with two inputs capital, $k$, and labour, $l$ with coefficients 0.1, 0.2, respectively.

Let $k$ be the only independent input, $log(k_0) \sim \mathcal{N}(10, 2)$.

All other parameters accept default values.

julia> using DLWGMMIV

julia> input_names = ["k","l"]

julia> num_inputs = length(input_names)

julia> prod_params = [0.1, 0.2]

julia> indp_inputs_lnmean = [10]

julia> indp_inputs_lnvariance = [2]

julia> df = DLWGMMIV.sim_data(20000, 10, 
    num_inputs = num_inputs,
    input_names = input_names, 
    prod_params = prod_params, 
    indp_inputs_lnmean = indp_inputs_lnmean, 
    indp_inputs_lnvariance = indp_inputs_lnvariance
    ).df;

Sim Data for 2 inputs, CD

k Parameters:
  k_prod_params = 0.1 | k_cost_params = 0.0
l Parameters:
  l_prod_params = 0.2 | l_cost_params = 0.15

  First order derivative at optimal L is approximately zero: true

  Second order derivative at optimal L check: true

=======================

SUMMARY:
        100.0% of observations passed first order conditions.
        100.0% of observations passed second order conditions.

=======================

Use the GMM IV to estimate the productivity parameters in industry A. Let's set bstart to [0.08, 0], assume a Cobb Douglas production function for industry A, and use both the Nelder Mead and LBFGS optimization methods.

julia> data = [df.time, df.firm, df.Y, df.k, df.l];

julia> bstart = [0.08, 0];

julia> res_nm = dlwGMMIV(data..., bstart = bstart);

julia> res_lbfgs = dlwGMMIV(data..., bstart = bstart, opt = "LBFGS");

julia> println("
Nelder Mead:\n
Converge = $(res_nm.conv_msg)\n
Objective Value = $(res_nm.other_results.crit)\n
betas = $(res_nm.beta_dlw)
\n
LBFGS:\n
Converge = $(res_lbfgs.conv_msg)\n
Objective Value = $(res_lbfgs.other_results.crit)\n
betas = $(res_lbfgs.beta_dlw)
")

Nelder Mead:

Converge = true

Objective Value = 4.965822136980516e-10

betas = (beta_x1 = 0.09539780702143216, beta_x2 = 0.17476298604850368)

LBFGS:

Converge = true

Objective Value = 1.42422490360313e-19

betas = (beta_x1 = 0.09539780702934003, beta_x2 = 0.17476298621274525)

TransLog Example

Now let's simulate data for industry B with 200 firms over 5 periods with an additional initial period 0. The production function in industry B is TransLog with three inputs capital, $k$, labour, $l$, and material, $m$ with coefficients 0.1, 0.25, 0.2, respectively, and for second order term $km$ the coefficient is 0.01.

All other parameters accept default values.

julia> using DLWGMMIV

julia> input_names = ["k","l","m"]

julia> num_inputs = length(input_names)

julia> prod_params = [0.1, 0.25, 0.2, 0, 0.01]

julia> df = DLWGMMIV.sim_data(200, 5, 
    num_inputs = num_inputs,
    input_names = input_names, 
    prod_params = prod_params,
    prodF = "tl" 
    ).df;

Sim Data for 3 inputs, tl

k Parameters:
  k_prod_params = 0.1 | k_cost_params = 0.0
l Parameters:
  l_prod_params = 0.25 | l_cost_params = 0.15
m Parameters:
  m_prod_params = 0.2 | m_cost_params = 0.0
kl Parameters:
  kl_prod_params = 0.0 |
km Parameters:
  km_prod_params = 0.01 |
lm Parameters:
  lm_prod_params = 0.0 |
k2 Parameters:
  k2_prod_params = 0.0 |
l2 Parameters:
  l2_prod_params = 0.0 |
m2 Parameters:
  m2_prod_params = 0.0 |

  First order derivative at optimal L is approximately zero: true

  Second order derivative at optimal L check: true

=======================

SUMMARY:
        100.0% of observations passed optimization generating the simulated data.
        100.0% of observations passed first order conditions.
        100.0% of observations passed second order conditions.

=======================

Use the GMM IV to estimate the productivity parameters in industry B. Let's set bstart to [0.08, 0, 0, 0, 0.015, 0, 0, 0, 0], assume a TransLog production function for industry B, and use both the Nelder Mead and LBFGS optimization methods.

julia> data = [df.time, df.firm, df.Y, df.k, df.l, df.m];

julia> bstart = [0.08, 0, 0, 0, 0.015, 0, 0, 0, 0];

julia> res_nm = dlwGMMIV(data..., bstart = bstart, prodF = "tl");

julia> res_lbfgs = dlwGMMIV(data..., bstart = bstart, opt = "LBFGS", prodF = "tl");

julia> println("
Nelder Mead:\n
Converge = $(res_nm.conv_msg)\n
Objective Value = $(res_nm.other_results.crit)\n
betas = $(res_nm.beta_dlw)
\n
LBFGS:\n
Converge = $(res_lbfgs.conv_msg)\n
Objective Value = $(res_lbfgs.other_results.crit)\n
betas = $(res_lbfgs.beta_dlw)
")

Nelder Mead:

Converge = false

Objective Value = 79.28554768851731

betas = (beta_x1 = 0.4453756727068648, beta_x2 = 1.2792439223699248, beta_x3 = -0.5006920829608563, beta_x1x2 = -0.21300433972699703, beta_x1x3 = 0.19994522566063533, beta_x2x3 = 0.0881743447457756, beta_x12 = -0.05639552788752075, beta_x22 = 0.08837596941633362, beta_x32 = -0.14734946617896608)

LBFGS:

Converge = false

Objective Value = 0.0015682997593657287

betas = (beta_x1 = 0.5759958463450249, beta_x2 = 13.674180582661819, beta_x3 = -9.992663746360366, beta_x1x2 = 4.593764648836302, beta_x1x3 = -4.5375950132058565, beta_x2x3 = -1.7111532397047498, beta_x12 = 0.11862766042073726, beta_x22 = -7.3662414691120714, beta_x32 = 7.057850742289901)